+++ crates/crypto/src/bls12_381/tests/sign_tests.rs
@@ -26,7 +26,7 @@ fn test_all_sign() {
         // run_dkg: full DKG ceremony
         |n, t| {
             let mut coordinator = DKGCoordinator::new(
-                |id: u32, threshold: usize, total_nodes: usize, session_id: u64, role| {
+                |id: u32, threshold: usize, total_nodes: usize, session_id: u128, role| {
                     <DKGNode as Dkg>::new(id, threshold, total_nodes, session_id, role)
                 },
                 n,
@@ -73,7 +73,7 @@ fn test_threshold_signing_different_share_subsets_same_signature() {
     let t = 3;
 
     let mut coordinator = DKGCoordinator::new(
-        |id: u32, threshold: usize, total_nodes: usize, session_id: u64, role| {
+        |id: u32, threshold: usize, total_nodes: usize, session_id: u128, role| {
             <DKGNode as Dkg>::new(id, threshold, total_nodes, session_id, role)
         },
         n,
