+++ crates/crypto/src/decaf377/tests/sign_tests.rs
@@ -23,7 +23,7 @@ fn test_all_sign() {
         // run_dkg: full DKG ceremony
         |n, t| {
             let mut coordinator = DKGCoordinator::new(
-                |id: u32, threshold: usize, total_nodes: usize, session_id: u64, role| {
+                |id: u32, threshold: usize, total_nodes: usize, session_id: u128, role| {
                     <DKGNode as Dkg>::new(id, threshold, total_nodes, session_id, role)
                 },
                 n,
@@ -57,7 +57,7 @@ fn test_frost_rejects_tampered_commitment_from_coordinator() {
     let t = 2;
 
     let mut coordinator = DKGCoordinator::new(
-        |id: u32, threshold: usize, total_nodes: usize, session_id: u64, role| {
+        |id: u32, threshold: usize, total_nodes: usize, session_id: u128, role| {
             <DKGNode as Dkg>::new(id, threshold, total_nodes, session_id, role)
         },
         n,
